% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_buildKernel.R
\name{bandwidth_select}
\alias{bandwidth_select}
\title{Select bandwidth in Gaussian kernel.}
\usage{
bandwidth_select(expr, method)
}
\arguments{
\item{expr}{A m gene by n location matrix of normalized gene expression matrix.}

\item{method}{The method used in bandwidth selection, "SJ" usually for small sample size data, "Silverman" usually for large sample size data.}
}
\value{
A numeric value of calculated bandwidth.
}
\description{
This function selects bandwidth in Gaussian kernel.
}
