% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA.R
\name{CreateSpatialPCAObject}
\alias{CreateSpatialPCAObject}
\title{Create the SpatialPCA object with filtering and normalization step.}
\usage{
CreateSpatialPCAObject(
  counts,
  location,
  covariate = NULL,
  project = "SpatialPCA",
  gene.type = "hvg",
  sparkversion = "spark",
  numCores_spark = 1,
  gene.number = NULL,
  customGenelist = NULL,
  min.loctions = 20,
  min.features = 20
)
}
\arguments{
\item{counts:}{Gene expression count matrix (matrix), m x n, m is the number of genes and n is the number of locations.}

\item{location:}{Spatial location matrix (matrix), n x d, d is dimensin of spatial coordinates, e.g. d=2 for locations on 2D space.}

\item{covariate:}{The covariates in experiments (data.frame, if any covariate included), n x q, n is the number of locations, q is the number of covariates.}

\item{project:}{Project names.}

\item{min.loctions:}{The features (genes) detected in at least this many loctions.}

\item{min.features:}{The locations where at least this many features (genes) are detected.}

\item{gene.type:}{The type of genes to be used: "hvg" for highly variable genes; "spatial" for spatially expressed genes; "custom" for user specified genes.}

\item{gene.number:}{The number of top highly variable genes if gene.selection=="hvg" (all HVG genes if number is not specified);
number of top spatially expressed genes if gene.selection=="spatial" (all significant spatially expressed genes if number is not specified).}

\item{customGenelist:}{A list of user specified genes if  gene.type=="custom".}

\item{sparkversion:}{In spatial gene selection, specify "spark" for small sample size data, "sparkx" for large sample size data.}

\item{numCores_spark:}{In spatial gene selection, specify the number of cores in SPARK package when selecting spatial genes.}
}
\value{
Returns SpatialPCA object with filtered gene expression matrix.
}
\description{
Create the SpatialPCA object with filtering and normalization step.
}
