% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_utilties.R
\name{fx_CHAOS}
\alias{fx_CHAOS}
\title{Calculate CHAOS score to measure clustering performance.}
\usage{
fx_CHAOS(clusterlabel, location)
}
\arguments{
\item{clusterlabel}{Cluster labels.}

\item{location}{A n by k matrix of spatial locations.}
}
\value{
A numeric value for CHAOS score.
}
\description{
CHAOS score measures the spatial continuity of the detected spatial domains.
Lower CHAOS score indicates better spatial domian clustering performance.
}
