% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_utilties.R
\name{fx_PAS}
\alias{fx_PAS}
\title{Calculate PAS score to measure clustering performance.}
\usage{
fx_PAS(clusterlabel, location)
}
\arguments{
\item{clusterlabel}{Cluster labels.}

\item{location}{A n by k matrix of spatial locations.}
}
\value{
A numeric value for PAS score.
}
\description{
PAS score measures the randomness of the spots that located outside of the spatial region where it was clustered to.
Lower PAS score indicates better spatial domian clustering performance.
}
