% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_analysis.R
\name{get_NMF}
\alias{get_NMF}
\title{Obtain NMF low dimensional components.}
\usage{
get_NMF(expr, PCnum)
}
\arguments{
\item{expr:}{A m gene by n location of gene expression matrix.}

\item{PCnum:}{An integer of number of low dimensional components we want to extract.}
}
\value{
A d by n matrix of low dimensional components from NMF.
}
\description{
This function uses RunNMF from STUtility package to calculate NMF low dimensional components.
}
