% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_utilties.R
\name{get_PCA}
\alias{get_PCA}
\title{Obtain PCA low dimensional components for dimension reduction methods comparison.}
\usage{
get_PCA(object)
}
\arguments{
\item{object:}{SpatialPCA object.}
}
\value{
A d by n matrix of low dimensional components from PCA. d is number of low dimensional components, which is same as number of Spatial PCs in SpatialPCA. n is sample size.
}
\description{
Obtain PCA low dimensional components for dimension reduction methods comparison.
}
