% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA.R
\name{high_resolution}
\alias{high_resolution}
\title{High-resolution spatial map construction.}
\usage{
high_resolution(
  info,
  K,
  kernelpara,
  ED,
  est_log_tau,
  est_W,
  est_sigma0,
  est_Z,
  PCnum
)
}
\arguments{
\item{info:}{A n by k matrix, current location matrix.}

\item{K:}{A n by n matrix, current kernel matrix.}

\item{kernelpara:}{An integer that represents different kernels: 1 - Gaussian kernel, 2 - cauchy kernel, 3 - quadratic kernel. Here it should be consistent with the kernel that used in modeling original spatial PCs.}

\item{ED:}{A n by n Distance matrix.}

\item{est_log_tau:}{Estimated log tau value.}

\item{est_W:}{Estimated loading matrix W.}

\item{est_sigma0:}{Estimated sigma_0 value.}

\item{est_Z:}{A object from SpatialPCA_estimate_Z function.}

\item{PCnum:}{Number of predicted spatial PCs.}
}
\value{
A list of objects
\item{Z_star}{Predicted Z matrix on new locations} 
\item{Location_star}{Coordinates of new locations}
}
\description{
This function predicts spatial PC values at new spatial locations.
}
