% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_buildKernel.R
\name{kernel_build}
\alias{kernel_build}
\title{Build kernel matrix.}
\usage{
kernel_build(kerneltype = "gaussian", location, bandwidth)
}
\arguments{
\item{kerneltype:}{The type of kernel to be used, either "gaussian", or "cauchy" for cauchy kernel, or "quadratic" for rational quadratic kernel.}

\item{location:}{Cell/spot corrdinates to compute the kernel matrix}

\item{bandwidth:}{A numeric value of bandwidth.}
}
\value{
The kernel matrix for spatial relationship between locations
}
\description{
This function calculates kernel matrix.
}
