% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_utilties.R
\name{louvain_clustering}
\alias{louvain_clustering}
\title{Obtain louvain clustering cluster labels.}
\usage{
louvain_clustering(knearest, latent_dat)
}
\arguments{
\item{knearest:}{A vector of integers, number of nearest neighbors for KNN graph construction in louvain clustering.}

\item{latent_dat:}{A d by n matrix of low dimensional components.}
}
\value{
A list of clustering results.
\item{cluster_label}{a list of cluster labels, each corresponds to the vector of nearest neighbors.}
\item{cluster_num}{a list of total number of clusters in each cluster label.}
}
\description{
This function performs louvain clustering on input low dimensional components.
}
