% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_analysis.R
\name{plot_factor_value}
\alias{plot_factor_value}
\title{Obtain factor value plot.}
\usage{
plot_factor_value(info, latent_dat, textmethod, pointsize = 2, textsize = 15)
}
\arguments{
\item{info:}{A n cell by k dimension of location matrix. n is cell number.}

\item{latent_dat:}{A d by n matrix of low dimensional components.}

\item{pointsize:}{The point size of each cell.}

\item{textsize:}{The text size in the legend.}
}
\value{
A list of ggplot objects for factor value plots.
}
\description{
This function visualizes the low dimensional component values.
}
