% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_utilties.R
\name{plot_RGB_tSNE}
\alias{plot_RGB_tSNE}
\title{Visualize RGB plot from tSNE.}
\usage{
plot_RGB_tSNE(location, latent_dat, pointsize = 2, textsize = 15)
}
\arguments{
\item{location}{A n by k location matrix. n is spot number.}

\item{latent_dat}{A d by n matrix of low dimensional components.}

\item{pointsize}{The point size of each spot.}

\item{textsize}{The text size in the legend.}
}
\value{
A list.
\item{RGB}{A data frame with five columns: x coordinate, y coordinate, R, G, and B color index}
\item{figure}{A ggplot object for RGB plot from tSNE}
}
\description{
We summarized the inferred low dimensional components into three tSNE components and visualized the three resulting components with red/green/blue (RGB) colors in the RGB plot.
}
