% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_utilties.R
\name{plot_RGB_UMAP}
\alias{plot_RGB_UMAP}
\title{Obtain RGB plot from UMAP.}
\usage{
plot_RGB_UMAP(info, latent_dat, pointsize = 2, textsize = 15)
}
\arguments{
\item{info:}{A n cell by k dimension of location matrix. n is cell number.}

\item{latent_dat:}{A d by n matrix of low dimensional components.}

\item{pointsize:}{The point size of each cell.}

\item{textsize:}{The text size in the legend.}
}
\value{
A list of objects.
\item{RGB}{A data frame with five columns: x coordinate, y coordinate, R, G, and B color index}
\item{figure}{A ggplot object for RGB plot from UMAP.}
}
\description{
We summarized the inferred low dimensional components into three UMAP. components and visualized the three resulting components with red/green/blue (RGB) colors in the RGB plot.
}
