% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_utilties.R
\name{refine_cluster_10x}
\alias{refine_cluster_10x}
\title{Refine clustering results for 10x ST or Visium data.}
\usage{
refine_cluster_10x(clusterlabels, location, shape = "square")
}
\arguments{
\item{clusterlabels}{The cluster label obtained (e.g. from louvain method or walktrap method).}

\item{location}{A n by 2 location matrix of spots.}

\item{shape}{Select shape='hexagon' for Visium data, 'square' for ST data.}
}
\value{
The refined cluster labels.
}
\description{
This function refines spatial clustering of ST or Visium data.
}
