% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_func.R
\name{scale_expr}
\alias{scale_expr}
\title{Scale expression values of each gene}
\usage{
scale_expr(expr)
}
\arguments{
\item{expr:}{A m gene by n location of normalized gene expression matrix.}
}
\value{
A m gene by n location of gene expression matrix with each row of gene scaled
}
\description{
This function scales the expression of each input gene.
}
