% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_buildKernel.R
\name{SpatialPCA_buildKernel}
\alias{SpatialPCA_buildKernel}
\title{Calculating kernel matrix from spatial locations.}
\usage{
SpatialPCA_buildKernel(
  object,
  kerneltype = "gaussian",
  bandwidthtype = "SJ",
  bandwidth.set.by.user = NULL,
  sparseKernel = FALSE,
  sparseKernel_tol = 1e-20,
  sparseKernel_ncore = 1
)
}
\arguments{
\item{object:}{SpatialPCA object.}

\item{kerneltype:}{The type of kernel to be used, either "gaussian", or "cauchy" for cauchy kernel, or "quadratic" for rational quadratic kernel.}

\item{bandwidthtype:}{The type of bandwidth to be used in Gaussian kernel, "SJ" for Sheather & Jones (1991) method (usually used in small size datasets), "Silverman" for Silverman's ‘rule of thumb’ method (1986)(usually used in large size datasets).}

\item{bandwidth.set.by.user:}{User could select their own bandwidth (a numeric value) if the recommended bandwidth doesn't work in their dataset.}

\item{sparseKernel:}{"TURE" if the user wants to use a sparse kernel matrix or "FALSE" if not. It is recommended to choose sparseKernel="TRUE" when sample size is very large.}

\item{sparseKernel_tol:}{When sparseKernel=TRUE, the cut-off value when building sparse kernel matrix, any element in the kernel matrix greater than sparseKernel_tol will be kept, otherwise will be set to 0 to save memory.}

\item{sparseKernel_ncore:}{When sparseKernel=TRUE, the number of CPU cores to build sparse kernel matrix.}
}
\description{
Calculating kernel matrix from spatial locations.
}
