% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA.R
\name{SpatialPCA_estimate_parameter}
\alias{SpatialPCA_estimate_parameter}
\title{Estimate parameters in SpatialPCA.}
\usage{
SpatialPCA_estimate_parameter(
  maxiter = 300,
  log_tau_ini = 0,
  dat_input,
  PCnum = 20
)
}
\arguments{
\item{maxiter:}{Maximum iteration number.}

\item{log_tau_ini:}{Initial value of log(tau). Because we want to make sure tau is greater than 0 during iterations.}

\item{dat_input:}{Object obtained from data_prepare_func function.}

\item{PCnum:}{Number of spatial PCs.}
}
\value{
A list of objects
\item{par}{estimated log(tau) value}
\item{value}{-log likelihood corresponding to estimated log tau value}
\item{counts}{number of calls to the SpatialPCA_estimate function}
\item{convergence}{An integer code. 0 indicates successful completion, 1 indicates iteration limit maxiter had been reached.}
\item{message}{A character string giving any additional information returned by the optimizer}
\item{T_m_trend}{Total running time for this function}
}
\description{
This function estimates parameter tau with optim function.
}
