% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA.R
\name{SpatialPCA_estimate_paras_largedata}
\alias{SpatialPCA_estimate_paras_largedata}
\title{Calculate -log likelihood value in parameter estimation in large sample size.}
\usage{
SpatialPCA_estimate_paras_largedata(param_ini, dat_input, PCnum = 20)
}
\arguments{
\item{param_ini:}{Input log(tau) value.}

\item{dat_input:}{Object obtained from data_prepare_func function}

\item{PCnum:}{Number of spatial PCs.}
}
\value{
A numeric value of -log likelihood.
}
\description{
This function calculate -log likelihood value.
}
