% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_func.R
\name{SpatialPCA_estimate_W}
\alias{SpatialPCA_estimate_W}
\title{Estimate the loading matrix.}
\usage{
SpatialPCA_estimate_W(parameter, dat_input, PCnum = PCnum)
}
\arguments{
\item{parameter:}{The estimated log(tau) value.}

\item{dat_input:}{The input data object.}

\item{PCnum:}{Number of spatial PCs.}
}
\value{
A list of objects.
\item{W}{The estimated loading matrix W.}
\item{sigma2_0}{The estimated sigma^2_0 value.}
}
\description{
This function estimates loading matrix W in SpatialPCA.
}
