% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA.R
\name{SpatialPCA_estimate_W_largedata}
\alias{SpatialPCA_estimate_W_largedata}
\title{Estimate W matrix in SpatialPCA when sample size is large.}
\usage{
SpatialPCA_estimate_W_largedata(parameter, dat_input, PCnum = 20)
}
\arguments{
\item{parameter:}{Estimated log(tau) value.}

\item{dat_input:}{Object obtained from data_prepare_func function}

\item{PCnum:}{Number of spatial PCs.}
}
\value{
A list of objects, the first one is estimated W matrix under given parameter tau; the second one is estimated sigma_0 value.
}
\description{
This function estimates loading matrix W.
}
