% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_func.R
\name{SpatialPCA_estimate_Z}
\alias{SpatialPCA_estimate_Z}
\title{Estimate spatial PCs in SpatialPCA.}
\usage{
SpatialPCA_estimate_Z(parameter, dat_input, estW)
}
\arguments{
\item{parameter:}{The estimated log(tau) value.}

\item{dat_input:}{The input data object.}

\item{estW:}{The output from function SpatialPCA_estimate_W.}
}
\value{
A list of objects.
\item{Z_hat}{The estimated spatial PC matrix Z matrix.} 
\item{YM_mat_inv}{YM_mat_inv matrix, will be used in high-resolution prediction.}
}
\description{
This function estimates spatial PC matrix Z.
}
