% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA.R
\name{SpatialPCA_estimate_Z_largedata}
\alias{SpatialPCA_estimate_Z_largedata}
\title{Estimate Z matrix in SpatialPCA when sample size is large.}
\usage{
SpatialPCA_estimate_Z_largedata(parameter, dat_input, estW, PCnum = 20)
}
\arguments{
\item{parameter:}{Estimated log(tau) value.}

\item{dat_input:}{Object obtained from data_prepare_func function.}

\item{estW:}{Estimated W matrix.}

\item{PCnum:}{Number of spatial PCs.}
}
\value{
A list of objects
\item{Z_hat}{estimated Z matrix} 
\item{mat_inv}{mat_inv matrix, will be used in high-resolution prediction} 
\item{YM_mat_inv}{YM_mat_inv matrix, will be used in high-resolution prediction}
}
\description{
This function estimates spatial PC matrix Z.
}
