% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_highresolution.R
\name{SpatialPCA_expr_pred}
\alias{SpatialPCA_expr_pred}
\title{High-resolution gene expression prediction.}
\usage{
SpatialPCA_expr_pred(object)
}
\arguments{
\item{object}{SpatialPCA object with high resolution predicted spatial PCs.
Users can optionally provide new locations at the original location scale.}
}
\value{
Returns SpatialPCA object with predicted gene expression on new locations.
We can predict normalized gene expression for the genes in the object@normalized_expr matrix.
}
\description{
High-resolution gene expression prediction.
}
