% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_multiple_sample.R
\name{SpatialPCA_Multiple_Sample}
\alias{SpatialPCA_Multiple_Sample}
\title{Multiple sample SpatialPCA.
In this extension, we construct the covariance matrix for the latent factors in the form of a block diagonal matrix:
it consists of the kernel matrices constructed using the spatial location information within each dataset, with zero correlation for pairs of locations across datasets.
This way, the latent factors within each dataset are correlated a priori across spatial locations, while the latent factors across datasets are not correlated a priori.
Certainly, if one wants to model the a priori correlation between latent factors across datasets, due to, for example, their similarity in the features extracted from histology images, then one can also modify the kernel matrices by constructing them using features other than spatial location information.}
\usage{
SpatialPCA_Multiple_Sample(
  count_list,
  location_list,
  gene.type = "spatial",
  sparkversion = "spark",
  numCores_spark = 5,
  gene.number = 3000,
  customGenelist = NULL,
  min.loctions = 20,
  min.features = 20,
  bandwidth_common = 0.1
)
}
\arguments{
\item{count_list}{A list of g by n count matrix, g is gene number, n is location number.}

\item{location_list}{A list of n by d location matrix, n is location number, d is location dimension. The rownames of each location matrix should match with the colnames of its corresponding count matrix.}
}
\value{
Returns SpatialPCA object with estimated Spatial PCs on locations.
}
\description{
Multiple sample SpatialPCA.
In this extension, we construct the covariance matrix for the latent factors in the form of a block diagonal matrix:
it consists of the kernel matrices constructed using the spatial location information within each dataset, with zero correlation for pairs of locations across datasets.
This way, the latent factors within each dataset are correlated a priori across spatial locations, while the latent factors across datasets are not correlated a priori.
Certainly, if one wants to model the a priori correlation between latent factors across datasets, due to, for example, their similarity in the features extracted from histology images, then one can also modify the kernel matrices by constructing them using features other than spatial location information.
}
