% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_SpatialPCs.R
\name{SpatialPCA_SpatialPCs}
\alias{SpatialPCA_SpatialPCs}
\title{Calculating Spatial PCs (latent factor matrix Z).}
\usage{
SpatialPCA_SpatialPCs(object, fast = FALSE, eigenvecnum = NULL)
}
\arguments{
\item{object}{SpatialPCA object.}

\item{fast}{Select fast=TRUE if the user wants to use low-rank approximation on the kernel matrix to calculate the spatial PCs, otherwise select FALSE.}

\item{eigenvecnum}{When fast=TRUE, eigenvecnum is the number of top eigenvectors and eigenvalues to be used in low-rank approximation in the eigen decomposition step for kernel matrix.
The default is NULL, if specified, it is recommended that these top eigen values explain >=90\% of the variance.
In estimating spatial PCs, we need larger number of eigenvectors in kernel matrix for more accurate estimation.}
}
\value{
Returns SpatialPCA object with estimated Spatial PCs.
}
\description{
Calculating Spatial PCs (latent factor matrix Z).
}
