% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_utilties.R
\name{walktrap_clustering}
\alias{walktrap_clustering}
\title{Obtain clustering cluster labels through walktrap method.}
\usage{
walktrap_clustering(clusternum, latent_dat, knearest = 100)
}
\arguments{
\item{clusternum}{The desired number of clusters the user wants to obtain.}

\item{latent_dat}{A d by n matrix of low dimensional components, d is number of PCs, n is number of spots.}

\item{knearest}{An integers, number of nearest neighbors for SNN graph construction in walktrap clustering.}
}
\value{
The cluster labels.
}
\description{
This function performs walktrap clustering on input low dimensional components.
}
