% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA.R
\name{bandwidth_select}
\alias{bandwidth_select}
\title{Select bandwidth in Gaussian kernel}
\usage{
bandwidth_select(expr, info, method)
}
\arguments{
\item{expr:}{A m gene by n location of gene expression matrix.}

\item{info:}{A n cell by k dimension of location matrix. n is cell number, k is dimension}

\item{method:}{Method used in bandwidth selection, "SJ" for Sheather & Jones (1991) method, 
"Silverman" for Silverman's ‘rule of thumb’ method (1986), "Scott" for Scott (1992) method}
}
\value{
A scalar of calculated bandwidth.
}
\description{
This function selects bandwidth in Gaussian kernel.
}
