% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_func.R
\name{data_prepare_func}
\alias{data_prepare_func}
\title{Calculate and prepare data.}
\usage{
data_prepare_func(
  expr,
  info,
  covariate = NA,
  kerneltype = "gaussian",
  bandwidthtype = "SJ",
  fast = FALSE
)
}
\arguments{
\item{expr:}{A m gene by n location of gene expression matrix.}

\item{info:}{A n cell by k dimension of location matrix. n is cell number, k is dimension (k=2 when location is 2D; k=3 when location is 3D)}

\item{covariate:}{A covariate by n location matrix. If there is no covariate, the default is "NA".}

\item{kerneltype:}{An character string. It is the type of kernel to be used in SpatialPCA. Default is "gaussian", other options include "cauchy" for cauchy kernel and "quadratic" for rational quadratic kernel.}

\item{bandwidthtype:}{A character string representing method used in bandwidth selection: "SJ" for Sheather & Jones (1991) method, 
"Silverman" for Silverman's ‘rule of thumb’ method (1986), "Scott" for Scott (1992) method.}

\item{fast:}{A logical value, select "TRUE" to accrelerate the algorithm by performing low-rank approximation on the kernel matrix, otherwise "FALSE".}
}
\value{
A list of data matrices.
\item{YM}{YM matrix.}
\item{ED}{ED distance matrix.}
\item{ED2}{ED2 distance square matrix.}
\item{tr_YMY}{tr_YMY scalar.}
\item{M}{M matrix.}
\item{H}{H matrix.}
\item{n}{Sample size.}
\item{Y}{Expression matrix.}
\item{q}{Number of covariates, if no input covariate is avaliable then q=1 because we added an intercept in the function.}
\item{K}{Kernel matrix.}
\item{bandwidth}{Bandwidth in the kernel matrix.}
\item{YMt}{Transpose of YM matrix.}
\item{delta}{Eigenvalues of the kernel matrix.}
\item{U}{Eigenvectors of the kernel matrix.}
\item{kerneltype}{The type of kernel function used.}
\item{bandwidthtype}{The type of bandwidth selection method used.}
\item{location}{A n cell by k dimension of location matrix.}
}
\description{
This function calculates needed information for SpatialPCA.
}
