% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA.R
\name{data_prepare_func}
\alias{data_prepare_func}
\title{Calculate and prepare necessary matrices, such as projection matrix, distance matrix.}
\usage{
data_prepare_func(expr, info, covariate = NA)
}
\arguments{
\item{expr:}{A m gene by n location of gene expression matrix.}

\item{info:}{A n cell by k dimension of location matrix. n is cell number, k is dimension (k=2 when location is 2D; k=3 when location is 3D)}

\item{covariate:}{A covariate by n location matrix. If no location related covariate used, the default is "NA".}
}
\value{
A list of data matrices.
\item{YM}{YM matrix}
\item{ED}{ED distance matrix}
\item{ED2}{ED2 distance square matrix}
\item{tr_YMY}{tr_YMY scalar}
\item{M}{M matrix}
\item{H}{H matrix}
\item{n}{n matrix}
\item{Y}{expression matrix}
\item{q}{number of covariates, if no input covariate is avaliable then q=1 because we added an intercept in the function.}
}
\description{
This function calculates necessary objects before applying SpatialPCA.
}
