% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_utilties.R
\name{get_PCA}
\alias{get_PCA}
\title{Obtain PCA low dimensional components for dimension reduction methods comparison.}
\usage{
get_PCA(expr, PCnum)
}
\arguments{
\item{expr}{Normalized gene expression g by n matrix. g is gene number, n is sample size.}

\item{PCnum}{Number of PCs.}
}
\value{
A d by n matrix of low dimensional components from PCA. d is number of low dimensional components, n is sample size.
}
\description{
Obtain PCA low dimensional components for dimension reduction methods comparison.
}
