% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_analysis.R
\name{get_PCA}
\alias{get_PCA}
\title{Obtain PCA low dimensional components.}
\usage{
get_PCA(expr, PCnum)
}
\arguments{
\item{expr:}{A m gene by n location of gene expression matrix.}

\item{PCnum:}{An integer of number of low dimensional components we want to extract.}
}
\value{
A d by n matrix of low dimensional components from PCA.
}
\description{
This function uses SVD calculate PCA low dimensional components.
}
