% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_functions.R
\name{high_resolution}
\alias{high_resolution}
\title{High-resolution spatial map construction.}
\usage{
high_resolution(dat_input, PCnum)
}
\arguments{
\item{dat_input:}{The input data object.}

\item{PCnum:}{Number of spatial PCs.}
}
\value{
A list of objects.
\item{Z_star}{Predicted Z matrix on new locations.} 
\item{Location_star}{Coordinates of new locations.}
}
\description{
This function predicts spatial PC values at new spatial locations.
}
