% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_func.R
\name{kernel_build}
\alias{kernel_build}
\title{Build kernels.}
\usage{
kernel_build(kernelpara = "gaussian", ED2, beta)
}
\arguments{
\item{kernelpara:}{An character string that represents different kernels. Default is "gaussian", other options include "cauchy" for cauchy kernel and "quadratic" for rational quadratic kernel.}

\item{ED2:}{An n by n distance squared matrix.}

\item{beta:}{A numeric value of bandwidth.}
}
\value{
An n by n kernel matrix.
}
\description{
This function builds kernel matrix.
}
