% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA.R
\name{kernel_build}
\alias{kernel_build}
\title{Build kernels.}
\usage{
kernel_build(kernelpara = "gaussian", ED2, beta)
}
\arguments{
\item{kernelpara:}{An integer that represents different kernels: "gaussian", "cauchy", or "quadratic".}

\item{ED2:}{An n by n distance square matrix.}

\item{beta:}{A scalar of bandwidth value.}
}
\value{
An n by n kernel matrix.
}
\description{
This function selects bandwidth in Gaussian kernel.
}
