% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_buildKernel.R
\name{kernel_build}
\alias{kernel_build}
\title{Build kernel matrix.}
\usage{
kernel_build(kerneltype = "gaussian", location, bandwidth)
}
\arguments{
\item{kerneltype}{The type of kernel to be used, either "gaussian", or "cauchy" for cauchy kernel, or "quadratic" for rational quadratic kernel, and "delaunday" for gaussian kernel built with non-linear Delaunay triangulation based distance.}

\item{location}{A n by d matrix of cell/spot location coordinates.}

\item{bandwidth}{A numeric value of bandwidth.}
}
\value{
The kernel matrix for spatial relationship between locations.
}
\description{
This function calculates kernel matrix from spatial locations.
}
