% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_utilties.R
\name{louvain_clustering}
\alias{louvain_clustering}
\title{Obtain clustering cluster labels through louvain method.}
\usage{
louvain_clustering(clusternum, latent_dat, knearest = 100)
}
\arguments{
\item{clusternum}{The desired number of clusters the user wants to obtain.}

\item{latent_dat}{A d by n matrix of low dimensional components, d is number of PCs, n is number of spots.}

\item{knearest}{An integers, number of nearest neighbors for KNN graph construction in louvain clustering.}
}
\value{
The cluster labels.
}
\description{
This function performs louvain clustering on input low dimensional components.
}
