% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_analysis.R
\name{plot_cluster}
\alias{plot_cluster}
\title{Obtain cluster labels visualized on locations.}
\usage{
plot_cluster(
  loc_x,
  loc_y,
  clusterlabel_in,
  pointsize = 3,
  text_size = 15,
  title_in,
  color_in
)
}
\arguments{
\item{loc_x:}{A vector of x coordiantes.}

\item{loc_y:}{A vector of y coordinates.}

\item{clusterlabel_in:}{A vector of integers, the cluster labels for each cell.}

\item{pointsize:}{An integer, the point size of each cell.}

\item{textsize:}{An integer, the text size in the legend.}

\item{title_in:}{A character string, the title you want to display at the top of the figure.}

\item{color_in:}{A vector of colors for each cluster.}
}
\value{
A ggplot object.
}
\description{
This function visualizes cluster labels on locations.
}
