% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_utilties.R
\name{plot_factor_value}
\alias{plot_factor_value}
\title{Visualize PCs on their locations.}
\usage{
plot_factor_value(location, PCs, textmethod, pointsize = 2, textsize = 15)
}
\arguments{
\item{location}{A n cell by k dimension of location matrix. n is cell number, k=2 if the spots are on 2D space.}

\item{PCs}{A d by n matrix of low dimensional components.}

\item{textmethod}{A text string of the name of method used to extract latent factors, e.g. "SpatialPCA" or "PCA". It will be shown as the title of the figures.}

\item{pointsize}{The point size of each location for visualization.}

\item{textsize}{The text size in the figure legend.}
}
\value{
A list of ggplot objects for factor value plots.
}
\description{
This function visualizes the low dimensional component values.
}
