% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_utilties.R
\name{plot_trajectory}
\alias{plot_trajectory}
\title{Visualize pseudotimes on locations.}
\usage{
plot_trajectory(
  pseudotime,
  location,
  clusterlabels,
  gridnum,
  color_in,
  pointsize = 5,
  arrowlength = 0.2,
  arrowsize = 1,
  textsize = 22
)
}
\arguments{
\item{pseudotime}{A length n vector of pseudotime inferred from Slingshot.}

\item{location}{A n by 2 data frame of spot locations.}

\item{clusterlabels}{A vector of integers, the cluster labels for each spot}

\item{gridnum}{Number of grids that evenly segment the whole tissue section.}

\item{color_in}{A vector of character strings representing colors for each cluster.}

\item{pointsize}{An integer, the point size of each spot}

\item{arrowlength}{An integer, the length of arrows inside a grid between one spot with smallest pseudotime and largest pseudotime.}

\item{arrowsize}{An integer, the size of arrows inside a grid between one spot with smallest pseudotime and largest pseudotime.}

\item{textsize}{An integer, the size of text in the figure.}
}
\value{
A ggplot object.
\item{Pseudotime}{A ggplot object visualizing pseudotime on locations.}
\item{Arrowplot1}{A ggplot object for arrows pointing from smallest pseudotime and largest pseudotime in each grid.}
\item{Arrowplot2}{A ggplot object for arrows pointing from largest pseudotime and smallest pseudotime in each grid.}
\item{Arrowoverlay1}{A ggplot object for arrows pointing from smallest pseudotime and largest pseudotime in each grid, overlayed on clustering plot.}
\item{Arrowoverlay2}{A ggplot object for arrows pointing from largest pseudotime and smallest pseudotime in each grid, overlayed on clustering plot.}
}
\description{
This function visualizes pseudotimes on locations.
}
