% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_functions.R
\name{SpatialPCA_estimate}
\alias{SpatialPCA_estimate}
\title{Estimation of parameters.}
\usage{
SpatialPCA_estimate(param_ini, dat_input, PCnum = PCnum)
}
\arguments{
\item{param_ini:}{Input log(tau) value. Because we need tau to be positive, we calculate exp(log(tau)) inside of this function.}

\item{dat_input:}{Object created in the data_prepare_func function.}

\item{PCnum:}{Number of spatial PCs.}
}
\value{
A numeric value of -log likelihood.
}
\description{
This function calculate -log likelihood value.
}
