% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_functions.R
\name{SpatialPCA_estimate_Z}
\alias{SpatialPCA_estimate_Z}
\title{Estimate spatial PCs in SpatialPCA.}
\usage{
SpatialPCA_estimate_Z(
  parameter,
  dat_input,
  estW,
  PCnum,
  fast = FALSE,
  eigenvecnum = NA
)
}
\arguments{
\item{parameter:}{The estimated log(tau) value.}

\item{dat_input:}{The input data object.}

\item{estW:}{The output from function SpatialPCA_estimate_W.}

\item{PCnum:}{Number of spatial PCs.}

\item{fast:}{A logic value, TRUE if one wants to use approximation to quickly calculate the spatial PCs, otherwise select FALSE. It is recommended to set fast=TRUE when the sample size is large.}

\item{eigenvecnum:}{An integer, the number of top eigenvectors and eigenvalues to be used in low-rank approximation in the eigen decomposition step for kernel matrix. 
The default is NA, if specified, it is recommended that these top eigen values explain 90% of the variance.}
}
\value{
A list of objects.
\item{Z_hat}{The estimated spatial PC matrix Z matrix.} 
\item{U}{Eigenvectors used in K.} 
\item{delta}{Eigenvalues used in K.}
}
\description{
This function estimates spatial PC matrix Z.
}
