% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_func.R
\name{SpatialPCA_estimate_Z}
\alias{SpatialPCA_estimate_Z}
\title{Estimate spatial PCs in SpatialPCA.}
\usage{
SpatialPCA_estimate_Z(parameter, dat_input, estW, PCnum, fast = FALSE)
}
\arguments{
\item{parameter:}{The estimated log(tau) value.}

\item{dat_input:}{The input data object.}

\item{estW:}{The output from function SpatialPCA_estimate_W.}

\item{PCnum:}{Number of spatial PCs.}

\item{fast:}{A logic value, TRUE if one wants to use approximation to quickly calculate the spatial PCs, otherwise select FALSE. It is recommended to set fast=TRUE when the sample size is large.}
}
\value{
A list of objects.
\item{Z_hat}{The estimated spatial PC matrix Z matrix.} 
\item{YM_mat_inv}{YM_mat_inv matrix, will be used in high-resolution prediction.}
}
\description{
This function estimates spatial PC matrix Z.
}
