% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_highresolution.R
\name{SpatialPCA_highresolution}
\alias{SpatialPCA_highresolution}
\title{High-resolution spatial map construction.}
\usage{
SpatialPCA_highresolution(object)
}
\arguments{
\item{object:}{SpatialPCA object.}
}
\value{
Returns SpatialPCA object with estimated Spatial PCs.
}
\description{
High-resolution spatial map construction.
}
