% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_SpatialPCs.R
\name{SpatialPCA_SpatialPCs}
\alias{SpatialPCA_SpatialPCs}
\title{Calculating Spatial PCs (latent factor matrix Z).}
\usage{
SpatialPCA_SpatialPCs(object, fast = FALSE, eigenvecnum = NULL)
}
\arguments{
\item{object:}{SpatialPCA object.}

\item{fast:}{A logic value, TRUE if one wants to use approximation to quickly calculate the spatial PCs, otherwise select FALSE. It is recommended to set fast=TRUE when the sample size is large.}

\item{eigenvecnum:}{An integer, the number of top eigenvectors and eigenvalues to be used in low-rank approximation in the eigen decomposition step for kernel matrix.
The default is NULL, if specified, it is recommended that these top eigen values explain 90\% of the variance. In estimating Z, we need larger number of eigenvectors in kernel matrix for accurate estimation.}
}
\value{
Returns SpatialPCA object with estimated Spatial PCs.
}
\description{
Calculating Spatial PCs (latent factor matrix Z).
}
