% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCA_analysis.R
\name{walktrap_clustering}
\alias{walktrap_clustering}
\title{Obtain walktrap clustering cluster labels.}
\usage{
walktrap_clustering(knearest, latent_dat)
}
\arguments{
\item{knearest:}{A vector of integers, number of nearest neighbors for SNN graph construction in walktrap clustering.}

\item{latent_dat:}{A d by n matrix of low dimensional components.}
}
\value{
A list of clustering results.
\item{cluster_label}{a list of cluster labels, each corresponding to the vector of nearest neighbors}
\item{cluster_num}{a list of total number of clusters, each corresponding to the vector of nearest neighbors}
}
\description{
This function performs walktrap clustering on input low dimensional components.
}
